Modulo_sensor/requirements.txt
# Core dependencies
fastapi>=0.105.0
pydantic>=2.7.0
pydantic-settings>=2.10.1



uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
ipython==8.17.2

# Optional for enhanced shell
ipdb==0.13.13
paho-mqtt==2.1.0
